Update tn-aws.md
Ran through this guide after ~14 months and found some areas didn't work and some were out of date.

Changes:
bsdec2-image-upload has been updated from 1.2.2 to 1.3.1. The new version is required to avoid SSL errors on upload. 1.3.1 is listed on freshports but not in the FreeBSD ports tree. Included instructions on how to build, as well as a manual patch.
KEY.pem needs to be created manually as IAM keys are csv.
Included details about which user permissions are needed as well as S3 bucket settings.
The two bhyve commands failed for me. Included alternate plan that combines the two and uses a shell script that comes with bhyve.
Expanded on the required network changes for the EC2 instance to be reachable.