Disabling SMB1 and mention WS-Discovery
Not everyone looks at advisory notes; the requirements for SMB shares to appear in Windows Explorer are often very confused and misinformation abounds, and it's a hugely common topic in help requests/forum posts ("If I disable it how can I make my shares appear?"). So lets mention it briefly here. Bound to be a "big deal" info for many users.