Update _index.md
I did my best to be informative but not too verbose since I think the Hub is really user-friendly and easy to navigate. The links gave trouble without being more specific about their locations (maybe because I was on a local build of the site), so let me know if they need to be changed for the live site.