DOCS-2509 SCALE article: 2FA
Add a new article to https://www.truenas.com/docs/scale/credentials/ that describes the purpose and functionality of Two-Factor Authentication. The article needs to be weighted according to the list of options in the SCALE Credentials menu.

Document the configuration options on the 2FA screen and include a full how-to that shows setting up 2FA, logging in to the UI when 2FA is enabled, and logging in to the shell (SSH) when 2FA is enabled.

I also reviewed the 2FAFields.md.part file.