Docs-3197 Update Support Article and Snippets
Fix grammar and style issues in the Support article and the linked snippets for the support table, adding a debug and making a suggestion.
Moved the Contacting Support snippet to the Enterprise tab, added a statement about access to Support Services requiring a contract to both the general support tab and the contacting support snippet.