Docs-3570 iSCSI Articles Part 1
This commit reorganizes the original iSCSIShare.md article into the following:
iSCSIShare.md is the new UI ref article in /CORE/UIReference/Sharing/iSCSI/
AddingiSCSIShare.md, UsingiSCSIShare.md, IncreasingiSCSIAvailableStorage.md are all in /CORE/CORETutorials. Trying to add a folder under tutorials resulted in a lot of broken links.
This also updates the _index.md and iSCSIRef.md snippet, but the links to this new 13.0 branch files don't work and return page not found errors.
It also adds a few new images to the 12.0 folder because creating the 13.0 images folder also results is broken links.