Docs-3562 Update Services LLDP Article
This commit creates a new how-to article ConfiguringLLDP.md with the alias from the old article.
Renames the LLDP.md article to LLDPScreen.md
Updates the old snippet and saves a new one, but the new is not used in either article. Snippets updated are ServicesLLDPFields.md.part and ServicesLLDPFields.md.
Local build and test, all links work.