Docs-3572 WebDAV Article Updates
This commit retitles the AFPShares.md and NFPShare.md articles so they display correctly in the CORETutorials navigation.
It moves the WebDAVShare.md to CORETutorials as it is all how-to content
Creates new WebDAVShareScreen.md and services WebDAVScreen.md articles in UIReference folders.
It updates the old snippets and corrects style and grammar issues.
Local build tests all links working.