Docs-3943 TC 2.2 Update SAML Article and Fix Image Links
This commit splits the current SAML.md article into to separate articles SAMLAD.md for the Active Directory procedure and SAMLGA for the Google Admin procedure.
It fixes issue with image links that were not displaying, and fixes the flow and headers to present the information better and make the content more readable.