DOCS-3924: Updating adding a new field
Ensure Display Device was added in the GPU section of the VM Wizard:
When checked it will ensure that the guest always has access to a video device. For headless installations like ubuntu server this is required for the guest to operate properly. However for cases where consumer would like to use GPU passthrough and does not want a display device added should uncheck this.