Docs-3291 SCALE 22.02 Reorg Storage Articles
This commit covers corrections to storage/disks/ articles (fix relref links), and reorganizes the rest of the storage articles. It adds new how-to and UI ref articles, moves permissions, fusion pools and encryption to the tutorial area. Permissions and Encryption article changes are minor in this PR and each will be updated and reorganized in separate PRs/branches. SLOG article not created yet as it required additional research.
This PR creates a few new snippet files and updates images to 22.02.