Docs-4168 Update and Create Backup Credential Articles
This commit adds new UI ref articles for  CloudCredentialsScreens.md and SSHCredentialScreens.md.
It adds new tutorial articles to add cloud credentials and configure SSH connections and keypairs.
The Credentials _index.md is not updated with the children shortcode yet as it doesn't work as desired yet so it will get added when I finish the rest of the credentials articles.