Docs-4187 SCALE 22.02 Updated and Add New Cloud Sync Articles
This commit adds new UI reference article  CloudSyncTasksScreens.md.
Moves the cloudSyncTask folder to tutorials and renames it to remove the spaces between words. Moves the how-to content out of the _index.md file to AddCloudSyncTasks.md. Adds snippets for how-to content shared between how-to articles. Cleans up the CloudSyncTaskGoogleDrive.md article to remove the CORE screens and content that does not apply in SCALE, creates two sections for the two methods of creating a cloud credential.
Fixes broken links in other articles related to cloud sync tasks and cloud credentials.