Docs-3449 SCALE 22.02 Update Local User Article Content
This commit creates the UI reference article content and renames the previous LocalUsersSCALE.md to LocalUserScreensSCALE.md
And updates the how-to content in the renamed LocalUsersSCALE.md to ManageLocalUsersScale.md article. It updates the Shell types to reflect those listed in the UI.
It also adds a hint warning about root user login going away in a future release.