Docs-4154 Update Local Group Articles
This commit renames the LocalGroup.md articles to LocalGroupsScreens.md and ManageLocalGroups.md
Updates the content and screenshots in the UI ref article to cover the screens and settings, and in the how-to to reflect screen changes.
It fixes broken links by renaming the LocalUser.md articles to what they changed to in PR1510.