Docs-3390 SCALE 22.02 Create Services Articles
To minimize the review process, this commit creates a folder/expand for Services in both UI Ref and Tutorials.
Moves the services content from the _index.md file to individual files, and removes tabs, and adds tags and taglist shortcodes, but no other content changes made.
Tutorial Services articles are exact duplicates of the UI ref articles but has one for SFTP that is not in UI Ref as there is no SFTP screen.
This PR just gets the files in place. Subsequent PRs will address content changes for each Service.