Docs-2940 SCALE 22.02 Updated Replication Articles
This commit deletes the old replication article as all its content moved into new how-to articles so it is not needed.
Adds the final changes to the new UI ref content in the ReplicationScreensSCALE.md article.
Adds the taglist shortcode to all how-to articles and fixes a few minor style issues.