Docs-3771 Update Article Content and Remove Duplicate
This commit creates the UI ref content for LaunchDockerImageScreens.md and the AppsScreensSCALE.md, it updates out of date content in tutorial articles, changes docker.md to a Pi-Hole tutorial, removes duplicated content in UI Ref with aliases added to tutorials that pointed to the UI ref URLs. Next commit updates Chia and UsingCatalogs.md as the images and some instructions are now incorrect.