Docs-4366 Update Quotas Articles Add Snippets
This commit updates the QuotaScreens.md UI ref article content for UI changes in Bluefin.
It adds two new snippets, one for  UserQuotasScreenSCALE.md the other for GroupQuotasScreenSCALE.md. and uses them in both the ManageQuotas.md and QuotasScreens.md articles. It adds new screen shots for the UI changes.
It removes references to bulk changes and UI elements removed or implemented in a different way.