Docs-3394 Update Disks Article
This commit updates the content in the UI ref article for DisksScreensSCALE to the new UI screens and functions.
It updates all the tutorial articles for disks to the new UI and functions.
It updates tags in SMART test articles in both the UI and tutorials. It deletes the DatasetsSCALE.md article that was old content and not removed in a previous PR.