Docs-3349 Update Cloud Sync Tasks Articles
This commit updates the UI and tutorial articles.
It removes content about creating a cloud credential from the cloud sync task, this option was removed in Bluefin but expected to return in a later release.
It has a broken link that I'm not able to fix because the PR with the change has not merged into Bluefin docs yet