Docs-4410 Update Nextcloud and Admin User Articles
This commit removes the non-media Nextcloud article that essentially duplicates the InstallNextCloudMedia.md article content, adds an alias for the other article to the media article. It updates the ManageLocalUsersSCALE.md article to include information about the admin user requirement to set Permit Sudo, and fixes a few things in the rootless login article content.