Update PoolsScreens.md
Added an example to the Expand Pool section: there have been multiple posts on the forums asking how to add a new single drive to single-drive striped pool so to create or re-create a mirror. Most forum posts on the topic are dated and refer to a CLI zpool attach procedure which is unnecessary since the expand function is available in the GUI, but the official documentation does not offer much in this area. Even reddit posts suggest a complete redeploy of the pool in this case, which is also unnecessary.

A simple reference to the example in this section of the docs could save a lot of time to most new users. A reference has also been added to the Pool Manager section, since it's were most users seem to get stuck in the process.