PD-31 Update Launch Docker Image Articles
This commit updates the UI ref LaunchDockerImageScreens.md article and the tutorial Docker.md article with the changes UI screens and configuration settings.
It adds the Delete app and Delete image dialogs to the AppsScreensSCALE.md article and in the UsingApps.md tutorial article.
It removes the Pi-hole install article using docker image and replaces it with an install article using the Install option on the pihole widget (the install wizard for the pihole app).