PD-40 Update Clustering Articles
This commit updates the snippets used in the SCALE clustering articles to fix issues with numbered lists.
It adds a brief introduction to the types of clustering currently documented in SCALE.
It moves the MinIO Clustering to a separate article, removes the Launch Docker Image content and updates the Install from widget instructions. It converts the _index.md article to an expand navigation article for MinIO.