PD-88 Update Cloud Credentials and Sync Task Articles
This commit updates the CloudCredentialScreens.md UI ref article to include the new default Storj, update the OpenStack Swift setting changes based on the Auth option selected (implemented with tabs for readability) and fixed issues with expands that hid content instead of including it in the TOC.
It also fixes the AddCloudCredential.md and AddCloudSyncTasks.md articles. The procedures were incomplete or incorrect. It removes obsolete content from both.