PD-80 Update Getting Started SetUpStorage.md Article
This commit updates the /gettingstarted/install/SetUpStorageScale.md article content.
It removes out of date content, inserts links to the current article content on setting up pools, the Storage Dashboard screen, setting up datasets and setting up zvols.
It adds a new heading for next steps and changes the link for data sharing to SCALE content and not CORE.