PD-112 Update Install.md Article
This commit updates the /install/install.md article content. It adds Enterprise customer information, fixes the install step 4 in both physical and vm installs to select option 1 to create the admin user.
It also adds a better summary at the bottom of the page for the next steps after the install.