PD-116 Update Core to Scale Migration Article
This commit updates the MigratingFromCORE.md article content by adding a section on preparing for migration, adding Enterprise content, adds content on when the CORE update function works to use it to migrate to SCALE. And it makes other style/organizational improvements.
When failover works properly, this article will include some information on migrating a non-Enterprise (HA) system from CORE to SCALE. This content is not included yet.