PD-108 Update Alerts and SMTP Content
This commit eliminates the tabs in the AlerManage.md article, adds the TOC, updates the information in the SMTP and email configuration section.
Adds the extra front matter parameters, adds tags and taglists, and the children shortcode to the index file.
This does not break the tutorials out into separate articles. That is deferred to the TC content reorg effort for later this year or when TC 3.0 is released.