PD-268 Update Set Up Storage Articles
This commit updates the /GettingStarted/Configure/SetUpStorageScale.md article with create pool and create dataset snippets, cleans up the list of tags.
It also updates the Storage/Pools/CreatePoolScale.md article with a new snippet, CreatePool.md.
It updates the /Storage/Datasets/DatasetsSCALE.md article with a new snippet, CreateDatasetSCALE.md.
This change applies to the Master branch as well as 22.12.2 so this should merge into Master.