PD-270 Update Backing Up Articles
This commit updates the /GettingStarted/Config/SetUpBackupSCALE.md article with new snippets for content shared in other tutorial articles. It changes the tags and taglist shortcodes to streamline the list of related articles.
It creates new snippets /content/_includes/CreateCloudSyncTaskSCALE.md and DownlooadSysConfiFileSCALE.md
It makes minor updates to other SCALETutorials articles to add the snippets, makes minor fixes and moves the AddReplicationSCALE.md article to the folder and navigation for /SCALETutorials/DataProtection/Replication/ folder.