PD-241 Update First Time Login Article
This commit updates the /GettingStarted/Configure/FirstTimeLogin.md article to serve as the introduction to the UI and the first step in the UI configuration process. Added snippets, reworked sections to point at other article rather than duplicate content, and to add enterprise shortcode for content that applies to just Enterprise customers.