PD-237 Make Migration Guide More Comprehensive
This commit reorganizes content originally found only in the MigratingFromCORE.md article and moves it into new snippets and articles.
It creates new content covering preparing for migration for all users, and one for "migrating" CORE HA to SCALE HA.
I moves the CLI commands list to a separate article.
The snippets are not working correctly in this commit.
It