PD-84 Update Replication Tutorials
This commit creates new  snippets and uses them across the three tutorials LocalReplicationSCALE.md, RemoteReplicationSCALE.md and the new ReplicationWithEncryptionSCALE.md article with a new procedure for doing a simple replication task for an encrypted dataset or pool.

This should probably go into 22.12.2_docs as there are changes only in the 12.2 release.