PD-283 Update ZFS Primer Article
This commit fixes the incorrect statement ti say ESXi over NFS, and reformats the article in general.
It corrects style issues, adds headings for ZFS features to make the content easier to find via the TOC.
It also makes updates to the content to make it active voice, present tense.
This content applies to CORE and SCALE so it merges into Master.