PD-324 Update MinIO Manual Update Article
This commit updates the MinioManualUpdate.md file adding a summary procedure linking the individual instruction sections to each step, and changes the section headers for the instructions to gerund form. It also adds the known workarounds for host path validation and turning off share service before deploying the app, then restarting the service after the app is running.