PD-324 Add Enterprise MinIO Article Content
This commit adds a new navigation section for Enterprise Applications and creates the /Apps/EnterpriseApps/_indext.md file.

It creates a new tutorial article for MinIO Enterprise users in the new section for Enterprise Apps.

It also moves the existing EnterpriseApplicationsSCALE.md article to the new section for Enterprise applications

It makes a few minor edits to the UI reference article.

Finally, it changes the navigation weights for the application tutorials to get them in the right order with the Enterprise Articles at the top of the list under Apps.