PD-380 Update Cluster Articles
This commit adds the new trueimage shortcode to clustering articles.

It adds information on supported cluster types, requirement to completely/properly remove gluster metadata from a server if if was in a cluster, before you can reuse it in a new cluster, and requirement that all TrueNAS systems must have the same release version.