PD-493 Change Migration Instructions
This commit/PR changes the OpenVPN migration instructions in the OpenVPNSCALE.md article to point to a general VPN application instead of WG-Easy.
It updates the InstallWGEasyApp.md migration instructions to refer to a general VPN application rather than the WG-Easy application. It makes the WG-Easy article more generic rather than being the VPN replacement article.