PD-249 Create Filesystem CLI Articles
This commit updates the SCALECLIReference/Filesystem/_index.md file content and makes it visible on the website.

It creates the CLIAcltemplate.md article for the filesystem acltemplate child namespace with most but not all commands documented. It makes this article visible, and includes work in progress labels in the command sections where further validation is required.