PD-371 Update MinIO App Index File
This commit moves the content from the InstallMinIOS3.md article into the MinIOApp/_index.md file and deleted the separate tutorial article.
It changes the Adding the MinIO Enterprise App section to an expand and changes the image shortcodes to the version without figure numbers.
It fixes broken relrefs links in other articles.