PD-615 Update User Articles
This commit makes CJTF recommended updates to the ManageLocalUserSCALE.md content and snippets. The article content now reads like a procedure and UI reference material.
Creates a new snippet RootToAdminUserAccoung.md to and adds it to the user account tutoral and the GettingStarted/Migrate/MigratingCoreToScale.md article to clarify when and why users must create the admin account.
Adds the new snippet to the MigratingFrom Core.md article.
It also updates the RootLoginWarnSCALE.md snippet.