PD-709 Add SMB Share Content to Syncthing Enterprise Article
This commit creates a new snippet for the first steps content in common with the Syncthing Charts and Enterprise apps and adds it to both tutorials.
It updates the Enterprise article content with the SMB share content, changes the instruction on opening the Syncthing app details screen, and makes general improvements to the content in general.