PD-774 CLI Guide Style Changes
This PR makes CLI guide style changes based on feedback received.

The Auth _index.md file is based on the Microsoft example with commands and descriptions listed in tables, and commands as links to the command details sections.
The Network _index.md file lists commands as links to the details sections with the descriptions of the command in hover text.

Command detail sections are reorganized to remove the extra expands, reword content where appropriate, remove the extra Description and Usage headings.

Web site work is planned to change how expand boxes display, and to make other suggested changes not included in these examples.