PD-847 Reorganize Getting Started Article Content
This commit updates the content of the articles in the new /TCGettingStarted section.

It renames the adminaccounts.md article to TCFirstTimeLogin.md and updates the content.
It updates the InitialConfig.md content Getting Started steps, removing creating admin account because that is already completed before you get to the instructions in this article. It updates screenshots and UI references and fixes style issues.
Updates the ConnectingTrueNAS.md article content to include references to all system screens with the New System function including the new fleet dashboard.
It updates UserAccounts.md content and instructions.
It updates the InterfaceOverview with new icons and changes to the navigation menus.