PD-825 Move CreatingBasicVM.md Content
This commit deletes the CreatingBasicVM.md article and moves that content to the /VirtualMachine/_index.md file.

It fixes issues with bullet list parallel construction and makes a few minor style and voice/tense changes.
It adds aliases moved from, and a new one for the removed CreatingBasicVM.md article

It fixes a broken article link in the /solutions/optimizations/security.md article.