PD-825 Move PoolCreate.md content
This commit deletes the PoolCreate.md article and moves the content to the /storage/pools/_index.md file.
It adds an alias to the deleted article and makes some minor changes to the content to improve readability.
It fixes broken links to the deleted PoolCreate.md article in other articles.