PD-825 Move PoolsScreen.md Content
This commit makes minor changes in the /storage/_index.md file to standardize the children shortcode section header.

It deletes the PoolsScreen.md file and moves the content into the /storage/pools/_index.md file.
It adds an alias for the deleted article and makes a minor change to the children shortcode section adding a heading for it.