PD-951 Update SettingUpJailStorage.md Article
This commit updates the content in the SettingUpJailStorage.md article to improve readability.

It adds new section headers and additional explanations about how to navigate to screens and what the user needs to do.
It adds iocage commands to add a new directory and mount point in an SSH session.
It adds new content for adding pointers to datasets from a plugin/jail directory.