PD-953 Update Advanced Settings Tutorials
This commit updates the /SystemSettings/Advanced/_index.md article, renames and updates the SessionsSettingsWidget.md snippet to the AccessSettingsWidget.md, and updates the ManageGlobal2faSCALE.md article content to reflect changes made in Dragonfish.

The Audit Widget and configuration screen is documented but a new article will cover this content under a different PR when the UI is ready to document.