PD-1011 Update CORE to SCALE Migration Articles
This PR updates the content in the SCALE/GettingStarted/Migrate/MigratePrep.md article by reorganizing and improving the content in the preparation steps. It adds additional information to the Deprecated Services section introduction and the S3 Service content.
It updates the MigratingFromCORE.md article by adding the missing step to upload the config file and verify the rest of the settings, beginning with network settings, then checking pools, and finally all other settings.
It deletes the MigrateCoreServicesToCobia.md snippet that was not used in other article (per search result) and moved the content into the preparation article.
Improves the content in MigrateCOREHAtoSCALEHA.md.
Updates the mermaid for the upgrade path in the SCALEReleaseNotes.md and adds information on the new force option for deprecated services.