PD-1010 Move UseAdvancedReplicationSCALE.md Content to Index.md
This commit moves the article content in UseAdvancedReplicationSCALE.md and moves it into the /AdvancedReplication/_index.md article, and deletes the UseAdvancedReplicationSCALE.md article.
It updates the _index.md content with an alias to the deleted file, and adds a hint admonition stating advanced replication requires an existing periodic snapshot task and matching naming schema in the snapshot and replication task or it fails.