PD-972 Update Screenshot, UI ref and Tutorial with ALUA Information
This commit updates the screenshot of the Global Target Configuration screen.
Updates the UI reference snippet to include the new screenshot and the table with info on this checkbox, caveated as visible for Enterprise-licensed system only.
It also updates the AddingiSCSIShares.md to include an optional step for this new checkbox, also caveated as Optional, only Enterprise Licensed systems.