PD-1011 PR Feedback Commit
This commit makes changes suggested in the PR review.

It creates a new snippet AddAdminUserAccount.md for the content originally in the ManageLocalUsersSCALE.md article and adds that snippet to that article and the Migrating from CORE article.

It changes the numbered sequence in the MigratePrep.md article to make the GELI step 2, and combines steps 6 and 7 into one on backing up data.