PD-917 Update ConsoleSetupMenueSCALE.md Article
This commit updates the content in this article by removing rootless login mentions and changes this to references to using administrator logins.
It makes style changes to mentions of Console setup menu to Console Setup menu to standardize on init caps of Console and Setup.
It makes other content changes to improve readability.