PD-917 Update images and content
This commit updates the FirstTimeLogin.md content to remove duplicated content for power button options and replaces it with links to the internal content found in the TopToolbar.md snippet. It makes other content changes.
It updates the TopToobar.md snippet content with new a screenshot and power option content related to admin roles.