PD-1000 Update Sharing Articles with Sharing Admin info
This commit creates a new snippet with information on the sharing admin and adds the shortcode for this to the adding SMB, NFS, and iSCSI share tutorial articles.
It updates the SMB Share/_index,md article content on editing the builtin-user ACE changing it to include instructions on when to modify or not modify it and how to add a new group ACE rather than modifying the builtin-user group.
It also updates the Windows instructions on mounting the share to add clarity on the process.