PD-397 Update Permissions Article Content
This PR updates the ACLPrimer.md reference article to include information on the NFSv4.1 ACL changes.

It updates the PermissionsSCALE.md article, adding in and reworking the content taken from the ManageSMBShares.md article, and adding a section on configuring/editing the POSIX ACL, and modifying the section on Configuring an NFSv4 ACL.

It changes header text on a few existing sections.