PD-817 Uodate Screenshots and Reporting Article Content
This PR updates the UI reference and tutorial articles for reporting.

Screenshots are updated to the current 24.04 screens.

The UI ref article is updated with new screens, removed the top table of screen options and moved that content down to be introductory sentences for each graph screen section.
Updated the Reporting Exporters with changes to the UI and information about the settings.

Updated the reporting tutorial content. Merged the two different sections on configuring exporters into one section, and added more information on configuring exporters. Screenshots also updated