PD-1019 Update Veeam Solutions Article
This commit adds a new screenshot of the Veeam Ready list of iX systems and a link to the Veeam website with the active links in the last column of the table.

It updates an incorrect reference to HA architecture being free, and changes Advantages section to mention the Enterprise hardware and not just hardware.

It changes the Performance Tuning for Veeam section from a bullet list to a numbered list and adds a list of required systems before the numbered steps.

It fixes typos found in the article.