PD-1071 Update main Apps index content
This PR updates the main Apps/_index.md file content, reorganizes the content into a better flow and adds new sections to explain the various screen options.
It adds general information on the typical setting sections found in most app installation wizards, and adds the new FileExplorerFolderIcons.md snippet to the section on Storage  Configuration.
It updates images to remove the "new" from the top banner in the screenshots.
These changes will be backported to the 24.04 branch in a separate PR.