Update SetUpBackupScale.md
Storj has ended their free 25GB tier as of 4/1/24, as well as the free 25GB of storage on pro accounts. Update any Storj content to reflect this change and determine if we want to go back to the old method of backing up storage since Storj is no longer the easiest (cheapest) way.