Update General Network Articles - 24.04
This PR updates content in the general network tutorial articles, creates a new snippet with HA network instructions and embeds that in the /GettingStarted/InstallEnterpriseHASCALE.md and ManageNetworkSettingsHA.md articles. It updates the AddingGlobalConf.md article content. Updates screenshots and changes image code to the trueimage shortcode. Updates content in a few snippets to remove duplicated text and repeat content in other articles that should also be included in other articles.