Update App Articles - 24.04
This PR updates the /apps/_index.md, apps/CommunityApps/_index.md, and apps/Enterprise/_index.md articles, and the UsingCatalogs.md article content covering catalogs, trains, and content that changed from Cobia to Dragonfish in the 24.04 branch. It removes references to docker and deletes the AppsAdvancedSettings.md article with host path information and settings that no longer apply.
These changes will be backported to Master in a separate PR.