Update Dataset UI Articles - 24.04
This PR updates the UI ref articles in the Dataset section:
/Datasets/_index.md
EditACLScreens.md
EncryptionUISCALE.md
QuotaScreens.md
SnapshotsScreens.md
ZvolsScreensSCALE.md
and makes minor updates to the CloneAndPromoteSnapshotDataset.md
Image shortcodes change to trueimage in a few articles, and added a new screenshot.
Updated content that was out of date, confusing, or incorrect.
These changes will be backported to Master in a different PR.