Udpdate UI Ref Article and Screenshots
This commit updates the VirtualizationScreens.md article content, reorganizes the information, changes the Add Device screens headings by device type to expands.
Adds the CloneVMDialog screenshot and updates the Edit screenshots to better images. It also converts the CPU And Memory, and GPU settings to snippets to use in the Wizard and Edit screen sections.