Fix issues with Links and procedure numbering
This commit fixes broken relref links to the deleted article, and replaces the Add bridge interface snippet with a link to the article. While not ideal, the numbering in the snippet resulted in confusion in the procedure in the AccessNASfromVM.md article because step 5 ended up with nested steps starting at 1.