Update Virtualization Articles - Master
This PR backports 24.04 changes to Master updates the Virtualization UI ref and tutorial article content.

It removes the CreatingManagingVMsSCALE.md article after moving the content to the Virtualization/_index.md file. The alias is added for the article removed.
It fixes issues with numbered procedures not rendering correctly by indenting the screenshots, and removing a snippet and replacing it with a link to fix the issue with the nested step numbering.
It updates and adds screenshots, and changes to the trueimage shortcode where necessary.