Update Cloud Sync Task Tutorials
This commit moves the content of the AddCloudSyncTasks.md article into the _index.md file and deletes the article. It updates links, fixes typos, corrects incorrect content.
It updates the CloudSyncTasksScreens.md article, adding new screenshots and content on the details screen, Cloudsync Task Wizard screens, and fixes out of date content.
It updates the cloud sync task snippets and other articles with the wizard information.