Update Rsync Task Articles - 24.04
This PR backports to 24.04 the changes made in the rsync task UI ref and tutorial articles in the master branch.

Updated screenshots added and compressed.

Removed duplicated text in the tutorial and added an expanded Before You Begin section covering preparation for each of the two methods.

Added missing UI ref content and clarified setting descriptions.