PD-1070 Update Resilvering Articles
This PR updates both the CORE and SCALE articles on resilvering priorities in Master.
It creates two new snippets for content shared in both CORE and SCALE.
It updates the SCALE Scrub Task UI ref with the Advanced Scheduler snippet/expand.
It makes minor changes to setting descriptions and instructions to clarify the information on setting begin and end priorities.
These changes can be backported into separate PRs, one for CORE 13.0 and the other for SCALE 24.04 and/or 23.10 as the content applies to those release versions.