PD-1005 Remove CLI Index and Procedure
This commit removes the main SCALECLIReference/_index.md file.

It also reworks the CLI procedure in the SLOG overprovisioning article for now to keep the disk resizing instructions until we have a path forward with engineering on disk resizing: use zfs/Linux commands or add capability to the UI via improvement ticket.