PD-1179 Update Hardware _index.md Articles With SEO
This PR adds the description and keyword front matter parameters to the hardware platform _index.md articles and adds SEO terms in both the description and the keywords.
It adds a link to the sales pages through the product image and a text-based link to the content.

It updates the layouts/shortcodes/prodintro.html file to allow the links from the image to work.
Do not backport these changes.