PD-1125 Update IPv6 Reference Article Content
This commit adds bullet point expands to condense content shown on-screen and allow readers to choose to read the content or skip over it.
It reworks content to remove duplication and unnecessary information.
It adds additional links to external resources.
It reorganizes the helpful commands section to group by the functional area of the commands.