PD-1203 Updates to Storj App Content
This commit changes list types in the Before you Begin and Installing Storj app sections.

It adds a note about the old port number assignment because there is enough content out in google searches about port 20988 instead of 28967 that this content should clear up any confusion on that point.
It also adds information about Storj multi-tenant solution architecture not being included. A user became confused about this requesting information on a datastore that is not covered in our article. This is not something the SCALE Storj app can be configured for as it is provided by Storj  to customers on the Storj site, not through the node.